web/requirements.txt
Django==1.11.7
django-redis==4.8.0
djangorestframework==3.7.3
gunicorn==19.7.1
psycopg2==2.7.3.2
pytz==2017.3
redis==2.10.6
celery==4.1.0
requests==2.18.4
djangorestframework-jwt==1.11.0
